handle location denial and location enabled cases
Following refusal of iOS’ initial request for location permissions,
this change introduces an alert when location is not available to the
app. The check occurs:
 1. when the canvasviewcontroller appears (i.e. when it is transitioned
    into)
 2. when the user denies setting the permissions
 3. when the user presses the Add Location button

There is an edge case where two alerts appear at the same time, and
that is when the app is newly installed and location services is
disabled. I opted to overlook this case because it is:
1) very unlikely
2) not a terrible experience
3) will only happen once