Fixes
Added fix to PersonDetailsViewController where the app was crashing
when the phone number is nil.

Also, fixed bug where we were cashing details like the phone number and
the voter’s email, even after they were deleted by the user.

Also, made changes to the UI so that, should the user add a mode of
communication, selects it as active and then deletes the info, the
relevant switches are deactivated. And, as well, if the data field is
empty we are unable to flick the switch to mark it as a viable mode of
communication.